authserver/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
PyJWT==2.3.0
python-dotenv==0.19.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
requests==2.31.0
